refactor(console): sie preview
