+++ packages/console/src/components/SignInExperiencePreview/types.ts
@@ -1,0 +1,5 @@
+export enum PreviewPlatform {
+  DesktopWeb = 'desktopWeb',
+  Mobile = 'mobile',
+  MobileWeb = 'mobileWeb',
+}
