+++ packages/console/src/pages/SignInExperience/components/Preview/index.tsx
@@ -1,21 +1,17 @@
 import type { LanguageTag } from '@logto/language-kit';
 import { languages as uiLanguageNameMapping } from '@logto/language-kit';
-import type { ConnectorResponse, ConnectorMetadata, SignInExperience } from '@logto/schemas';
-import { ConnectorType, AppearanceMode } from '@logto/schemas';
-import { conditional } from '@silverhand/essentials';
+import type { SignInExperience } from '@logto/schemas';
+import { AppearanceMode } from '@logto/schemas';
 import classNames from 'classnames';
-import { format } from 'date-fns';
-import { useEffect, useMemo, useState, useRef, useCallback, useContext } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import useSWR from 'swr';
 
-import PhoneInfo from '@/assets/images/phone-info.svg';
 import LivePreviewButton from '@/components/LivePreviewButton';
 import Select from '@/components/Select';
+import SignInExperiencePreview from '@/components/SignInExperiencePreview';
+import { PreviewPlatform } from '@/components/SignInExperiencePreview/types';
 import TabNav, { TabNavItem } from '@/components/TabNav';
 import ToggleThemeButton from '@/components/ToggleThemeButton';
-import { AppEndpointsContext } from '@/contexts/AppEndpointsProvider';
-import type { RequestError } from '@/hooks/use-api';
 import useUiLanguages from '@/hooks/use-ui-languages';
 
 import * as styles from './index.module.scss';
@@ -38,11 +34,8 @@ const Preview = ({
   const [mode, setMode] = useState<Omit<AppearanceMode, AppearanceMode.SyncWithSystem>>(
     AppearanceMode.LightMode
   );
-  const [platform, setPlatform] = useState<'desktopWeb' | 'mobile' | 'mobileWeb'>('desktopWeb');
-  const { data: allConnectors } = useSWR<ConnectorResponse[], RequestError>('api/connectors');
-  const previewRef = useRef<HTMLIFrameElement>(null);
-  const { customPhrases, languages } = useUiLanguages();
-  const { userEndpoint } = useContext(AppEndpointsContext);
+  const [platform, setPlatform] = useState<PreviewPlatform>(PreviewPlatform.DesktopWeb);
+  const { languages } = useUiLanguages();
 
   useEffect(() => {
     if (!signInExperience?.color.isDarkModeEnabled) {
@@ -74,64 +67,6 @@ const Preview = ({
     }
   }, [language, availableLanguageOptions]);
 
-  const config = useMemo(() => {
-    if (!allConnectors || !signInExperience) {
-      return;
-    }
-
-    const socialConnectors = signInExperience.socialSignInConnectorTargets.reduce<
-      Array<ConnectorMetadata & { id: string }>
-    >(
-      (previous, connectorTarget) => [
-        ...previous,
-        ...allConnectors.filter(({ target }) => target === connectorTarget),
-      ],
-      []
-    );
-
-    const hasEmailConnector = allConnectors.some(({ type }) => type === ConnectorType.Email);
-
-    const hasSmsConnector = allConnectors.some(({ type }) => type === ConnectorType.Sms);
-
-    return {
-      signInExperience: {
-        ...signInExperience,
-        socialConnectors,
-        forgotPassword: {
-          email: hasEmailConnector,
-          sms: hasSmsConnector,
-        },
-      },
-      language,
-      mode,
-      platform: platform === 'desktopWeb' ? 'web' : 'mobile',
-      isNative: platform === 'mobile',
-    };
-  }, [allConnectors, language, mode, platform, signInExperience]);
-
-  const postPreviewMessage = useCallback(() => {
-    if (!config || !customPhrases) {
-      return;
-    }
-
-    previewRef.current?.contentWindow?.postMessage(
-      { sender: 'ac_preview', config },
-      userEndpoint?.origin ?? ''
-    );
-  }, [userEndpoint?.origin, config, customPhrases]);
-
-  useEffect(() => {
-    postPreviewMessage();
-
-    const iframe = previewRef.current;
-
-    iframe?.addEventListener('load', postPreviewMessage);
-
-    return () => {
-      iframe?.removeEventListener('load', postPreviewMessage);
-    };
-  }, [postPreviewMessage]);
-
   return (
     <div className={classNames(styles.preview, className)}>
       <div className={styles.header}>
@@ -158,58 +93,36 @@ const Preview = ({
       </div>
       <TabNav className={styles.nav}>
         <TabNavItem
-          isActive={platform === 'desktopWeb'}
+          isActive={platform === PreviewPlatform.DesktopWeb}
           onClick={() => {
-            setPlatform('desktopWeb');
+            setPlatform(PreviewPlatform.DesktopWeb);
           }}
         >
           {t('sign_in_exp.preview.desktop_web')}
         </TabNavItem>
         <TabNavItem
-          isActive={platform === 'mobileWeb'}
+          isActive={platform === PreviewPlatform.MobileWeb}
           onClick={() => {
-            setPlatform('mobileWeb');
+            setPlatform(PreviewPlatform.MobileWeb);
           }}
         >
           {t('sign_in_exp.preview.mobile_web')}
         </TabNavItem>
         <TabNavItem
-          isActive={platform === 'mobile'}
+          isActive={platform === PreviewPlatform.Mobile}
           onClick={() => {
-            setPlatform('mobile');
+            setPlatform(PreviewPlatform.Mobile);
           }}
         >
           {t('sign_in_exp.preview.native')}
         </TabNavItem>
       </TabNav>
-      <div
-        className={classNames(styles.body, platform === 'desktopWeb' ? styles.web : styles.mobile)}
-        style={conditional(
-          platform === 'desktopWeb' && {
-            // Set background color to match iframe's background color on both dark and light mode.
-            backgroundColor: mode === AppearanceMode.DarkMode ? '#000' : '#e5e1ec',
-          }
-        )}
-      >
-        <div className={styles.deviceWrapper}>
-          <div className={classNames(styles.device, styles[String(mode)])}>
-            {platform !== 'desktopWeb' && (
-              <div className={styles.topBar}>
-                <div className={styles.time}>{format(Date.now(), 'HH:mm')}</div>
-                <PhoneInfo />
-              </div>
-            )}
-            <iframe
-              ref={previewRef}
-              // Allow all sandbox rules
-              sandbox={undefined}
-              src={new URL('/sign-in?preview=true', userEndpoint).toString()}
-              tabIndex={-1}
-              title={t('sign_in_exp.preview.title')}
-            />
-          </div>
-        </div>
-      </div>
+      <SignInExperiencePreview
+        platform={platform}
+        mode={mode}
+        language={language}
+        signInExperience={signInExperience}
+      />
     </div>
   );
 };
